refactor: extract BulkCollectionEditingAffix and CollectionTemplateRenderer components
Split the large inline JSX blocks in apps/frontend/app/components/common/index.tsx
into focused, reusable components to improve code organization and maintainability.

Changes:
- Extract BulkCollectionEditingAffix component (lines 370-555) to its own file
  - Contains modal/affix UI, selection buttons, and bulk collection logic
  - Explicit TypeScript interface: BulkCollectionEditingAffixProps
  - Handles target entities, loading states, confirmation dialogs

- Extract CollectionTemplateRenderer component (lines 556-626) to its own file
  - Renders individual template form fields based on CollectionExtraInformationLot
  - Explicit TypeScript interface: CollectionTemplateRendererProps
  - Supports String, Boolean, Number, Date, DateTime, StringArray field types

- Update main index.tsx to remove extracted components and clean unused imports
  - Remove FormEvent, DateInput, DateTimePicker, Affix, Paper, Switch, etc.
  - Remove CollectionExtraInformation, EntityToCollectionInput, Scalars types
  - Remove useState, useFormValidation, useBulkEditCollection hooks

- Update 8 importing files to use direct imports instead of re-exports:
  - CollectionTemplateRenderer: 2 form components
  - BulkCollectionEditingAffix: 6 route components

This refactoring improves separation of concerns, reduces file size (from 795 to
355 lines), and makes the components more testable and maintainable while
preserving all existing functionality.
